+++ .eslintrc.cjs
@@ -1,0 +1,22 @@
+/** @type {import("eslint").Linter.Config} */
+module.exports = {
+  root: true,
+  env: {
+    node: true,
+    es2022: true,
+  },
+  parser: "@typescript-eslint/parser",
+  parserOptions: {
+    ecmaVersion: "latest",
+    sourceType: "module",
+    project: "./tsconfig.json",
+    tsconfigRootDir: __dirname,
+  },
+  plugins: ["@typescript-eslint"],
+  extends: ["eslint:recommended", "plugin:@typescript-eslint/recommended"],
+  ignorePatterns: ["out/**", "node_modules/**"],
+  rules: {
+    quotes: ["error", "double", { avoidEscape: true }],
+    semi: ["error", "always"],
+  },
+};